Rename file components/module-settings/module-settings-form
