+++ _inc/client/security/antispam.jsx
@@ -22,7 +22,7 @@ import analytics from 'lib/analytics';
 import { FormFieldset, FormLabel } from 'components/forms';
 import {
 	withModuleSettingsFormHelpers as withModuleSettingsFormHelpers,
-} from 'components/module-settings/module-settings-form';
+} from 'components/module-settings/with-module-settings-form-helpers';
 import SettingsCard from 'components/settings-card';
 import SettingsGroup from 'components/settings-group';
