+++ _inc/client/security/backups-scan.jsx
@@ -16,7 +16,7 @@ import Banner from 'components/banner';
 import { FEATURE_SECURITY_SCANNING_JETPACK } from 'lib/plans/constants';
 import {
 	withModuleSettingsFormHelpers as withModuleSettingsFormHelpers,
-} from 'components/module-settings/module-settings-form';
+} from 'components/module-settings/with-module-settings-form-helpers';
 import SettingsCard from 'components/settings-card';
 import SettingsGroup from 'components/settings-group';
 import {
