+++ _inc/client/security/protect.jsx
@@ -17,7 +17,7 @@ import { FormFieldset, FormLegend, FormLabel } from 'components/forms';
 import { ModuleToggle } from 'components/module-toggle';
 import {
 	withModuleSettingsFormHelpers as withModuleSettingsFormHelpers,
-} from 'components/module-settings/module-settings-form';
+} from 'components/module-settings/with-module-settings-form-helpers';
 import SettingsCard from 'components/settings-card';
 import SettingsGroup from 'components/settings-group';
