+++ _inc/client/traffic/related-posts.jsx
@@ -15,7 +15,7 @@ import {
 	FormLabel
 } from 'components/forms';
 import { ModuleToggle } from 'components/module-toggle';
-import { withModuleSettingsFormHelpers as withModuleSettingsFormHelpers } from 'components/module-settings/module-settings-form';
+import { withModuleSettingsFormHelpers as withModuleSettingsFormHelpers } from 'components/module-settings/with-module-settings-form-helpers';
 import SettingsCard from 'components/settings-card';
 import SettingsGroup from 'components/settings-group';
