+++ _inc/client/traffic/search.jsx
@@ -10,7 +10,7 @@ import Card from 'components/card';
  * Internal dependencies
  */
 import { FEATURE_SEARCH_JETPACK } from 'lib/plans/constants';
-import { withModuleSettingsFormHelpers as withModuleSettingsFormHelpers } from 'components/module-settings/module-settings-form';
+import { withModuleSettingsFormHelpers as withModuleSettingsFormHelpers } from 'components/module-settings/with-module-settings-form-helpers';
 import { ModuleToggle } from 'components/module-toggle';
 import SettingsCard from 'components/settings-card';
 import SettingsGroup from 'components/settings-group';
