+++ _inc/client/writing/composing.jsx
@@ -18,7 +18,7 @@ import {
 } from 'components/forms';
 import { isModuleFound as _isModuleFound } from 'state/search';
 import { ModuleToggle } from 'components/module-toggle';
-import { withModuleSettingsFormHelpers as withModuleSettingsFormHelpers } from 'components/module-settings/module-settings-form';
+import { withModuleSettingsFormHelpers as withModuleSettingsFormHelpers } from 'components/module-settings/with-module-settings-form-helpers';
 import { getModule } from 'state/modules';
 import TagsInput from 'components/tags-input';
 import SettingsCard from 'components/settings-card';
