+++ _inc/client/writing/custom-content-types.jsx
@@ -10,7 +10,7 @@ import CompactFormToggle from 'components/form/form-toggle/compact';
  * Internal dependencies
  */
 import { FormFieldset } from 'components/forms';
-import { withModuleSettingsFormHelpers as withModuleSettingsFormHelpers } from 'components/module-settings/module-settings-form';
+import { withModuleSettingsFormHelpers as withModuleSettingsFormHelpers } from 'components/module-settings/with-module-settings-form-helpers';
 import { getModule, getModuleOverride } from 'state/modules';
 import { isModuleFound as _isModuleFound } from 'state/search';
 import SettingsCard from 'components/settings-card';
